+++ common/README.md
@@ -1,0 +1,3 @@
+# Common Modules
+
+This directory serves as an area for keeping commonly-used components, for re-use.
